pyproject.toml
[project]
name = "sharehold"
version = "0.1.0"
description = ""
authors = [
    {name = "Aron Sadegh ",email = "95709170+Aron-S-G-H@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django (==5.0)",
    "elasticsearch-dsl (==8.17.1)",
    "mssql-django (>=1.5,<2.0)",
    "django-elasticsearch-dsl (>=8.0,<9.0)",
    "pymemcache (>=4.0.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
